scripts/requirements.txt
clickhouse_driver~=0.2.7
wheel~=0.41.2
cryptography~=42.0.7
pillow~=10.3.0
Jinja2~=3.1.4
pytz~=2024.1
setuptools~=68.2.0
tzlocal~=5.2
typing_extensions~=4.11.0
orjson~=3.10.3
ujson~=5.9.0
pdfplumber~=0.11.0
uvicorn~=0.29.0
requests~=2.31.0
fastapi~=0.111.0
docx~=0.2.4
python-docx~=1.1.2
pip~=24.0
pdfkit~=1.0.0
aspose.slides~=24.4.0